Enhance MaxiBlock optimization
