+++ src/blocks/container-maxi/save.js
@@ -3,8 +3,7 @@
  */
 import { ArrowDisplayer, ShapeDivider } from '../../components';
 import { getGroupAttributes } from '../../extensions/styles';
-import MaxiBlock from '../../components/maxi-block';
-import { getMaxiBlockAttributes } from '../../extensions/maxi-block';
+import { MaxiBlock, getMaxiBlockAttributes } from '../../components/maxi-block';
 
 /**
  * External dependencies
