+++ src/blocks/image-maxi/edit.js
@@ -16,13 +16,8 @@ import {
 	getGroupAttributes,
 	getLastBreakpointAttribute,
 } from '../../extensions/styles';
-import MaxiBlock from '../../components/maxi-block';
-import {
-	getMaxiBlockAttributes,
-	MaxiBlockComponent,
-	withMaxiProps,
-} from '../../extensions/maxi-block';
-
+import { MaxiBlock, getMaxiBlockAttributes } from '../../components/maxi-block';
+import { MaxiBlockComponent, withMaxiProps } from '../../extensions/maxi-block';
 import {
 	BlockResizer,
 	Button,
