+++ src/blocks/number-counter-maxi/save.js
@@ -6,8 +6,7 @@ eslint-disable react/no-unknown-property
 /**
  * Internal dependencies
  */
-import MaxiBlock from '../../components/maxi-block';
-import { getMaxiBlockAttributes } from '../../extensions/maxi-block';
+import { MaxiBlock, getMaxiBlockAttributes } from '../../components/maxi-block';
 
 /**
  * External dependencies
