+++ src/blocks/svg-icon-maxi/save.js
@@ -6,8 +6,7 @@ import { RawHTML } from '@wordpress/element';
 /**
  * Internal dependencies
  */
-import MaxiBlock from '../../components/maxi-block';
-import { getMaxiBlockAttributes } from '../../extensions/maxi-block';
+import { MaxiBlock, getMaxiBlockAttributes } from '../../components/maxi-block';
 
 /**
  * Save
