+++ src/extensions/maxi-block/index.js
@@ -1,4 +1,3 @@
-export { default as getMaxiBlockAttributes } from './getMaxiBlockAttributes';
 export { default as handleSetAttributes } from './handleSetAttributes';
 export { default as MaxiBlockComponent } from './maxiBlockComponent';
 export { default as withMaxiProps } from './withMaxiProps';
